requirements.txt
actionlib==1.14.0
aiohttp==3.9.3
aiosignal==1.3.1
allennlp==2.10.1
allennlp-models==2.10.1
angles==1.9.13
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
base58==2.1.1
beautifulsoup4==4.12.3
blis==0.7.11
bondpy==1.8.6
boto3==1.34.64
botocore==1.34.64
cached-path==1.1.6
cachetools==5.3.3
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catalogue==2.0.10
catkin==0.8.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.0.4
cloudpathlib==0.16.0
commonmark==0.9.1
confection==0.1.4
conllu==4.4.2
controller-manager==0.20.0
controller-manager-msgs==0.20.0
crosslingual-coreference==0.2.3
cv-bridge==1.16.2
cymem==2.0.8
datasets==2.18.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dill==0.3.8
docker-pycreds==0.4.0
dynamic-reconfigure==1.7.3
en-core-web-sm==3.7.1
exceptiongroup==1.2.0
fairscale==0.4.6
filelock==3.7.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-core==2.4.1
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
h5py==3.10.0
huggingface-hub==0.21.4
idna==3.6
image-geometry==1.16.2
iniconfig==2.0.0
interactive-markers==1.12.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonnet==0.20.0
langcodes==3.3.0
laser-geometry==1.6.7
lmdb==1.4.1
MarkupSafe==2.1.5
message-filters==1.16.0
more-itertools==10.2.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
neo4j==5.18.0
nltk==3.8.1
numpy==1.24.4
packaging==24.0
pandas==2.0.3
pathlib-abc==0.1.1
pathtools==0.1.2
pathy==0.11.0
pillow==10.2.0
pipdeptree==2.16.1
pluggy==1.4.0
preshed==3.0.9
promise==2.3
protobuf==3.20.3
psutil==5.9.8
py-rouge==1.1
pyarrow==15.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic-core==2.16.3
pygments==2.17.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
pytz==2024.1
PyYAML==6.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
regex==2023.12.25
requests==2.31.0
resource-retriever==1.12.7
rich==12.6.0
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.12
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-launch==0.4.9
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.4
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rsa==4.9
rviz==1.14.20
s3transfer==0.10.1
sacremoses==0.1.1
scikit-learn==1.3.2
scipy==1.10.1
sensor-msgs==1.13.1
sentencepiece==0.2.0
sentry-sdk==1.42.0
setproctitle==1.3.3
shortuuid==1.0.13
six==1.16.0
smach==2.5.2
smach-ros==2.5.2
smart-open==6.4.0
smclib==1.8.6
smmap==5.0.1
soupsieve==2.5
spacy==3.7.4
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.1.5
srsly==2.4.8
tensorboardX==2.6.2.2
termcolor==1.1.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
thinc==8.2.3
think==0.1.2
threadpoolctl==3.3.0
tokenizers==0.15.2
tomli==2.0.1
topic-tools==1.16.0
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tqdm==4.66.2
traitlets==5.14.2
transformers==4.17.0
typer==0.4.2
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.12.21
wasabi==0.10.1
wcwidth==0.2.13
weasel==0.3.4
wikipedia==1.4.0
word2number==1.1
xacro==1.14.17
xxhash==3.4.1
yarl==1.9.4
